get items for shopcart

+++ OrdersAndShopCartAPI/Models/Product.cs
@@ -8,6 +8,7 @@ namespace OrdersAndShopCartAPI.Models
 
     public class Product
     {
+
         public int Id { get; set; }
         public string Name { get; set; }
         public string Brand { get; set; }
@@ -15,12 +16,12 @@ namespace OrdersAndShopCartAPI.Models
         public decimal? Price { get; set; }
         public int? RAM { get; set; }
         public int? Year { get; set; }
-        public float? Display { get; set; }
+        public double? Display { get; set; }
         public int? Battery { get; set; }
         public int? Camera { get; set; }
         public string Image { get; set; }
         public int? Quantity { get; set; }
-        public float? Memory { get; set; }
+        public double? Memory { get; set; }
         public string Color { get; set; }
         public string Description { get; set; }
     }
