+++ WebClient/Models/ProductModel.cs
@@ -7,6 +7,7 @@ namespace WebClient.Models
 {
     public class ProductModel
     {
+
         public int Id { get; set; }
         public string Name { get; set; }
         public string Brand { get; set; }
@@ -14,12 +15,12 @@ namespace WebClient.Models
         public decimal? Price { get; set; }
         public int? RAM { get; set; }
         public int? Year { get; set; }
-        public decimal? Display { get; set; }
+        public double? Display { get; set; }
         public int? Battery { get; set; }
         public int? Camera { get; set; }
         public string Image { get; set; }
         public int? Quantity { get; set; }
-        public decimal? Memory { get; set; }
+        public double? Memory { get; set; }
         public string Color { get; set; }
         public string Description { get; set; }
     }
